Create AppropriationTree component
